fix: removes ejs and renames sidebar links
