--- frontend/html/includes/_footer.ejs
@@ -1,106 +1,0 @@
-<footer class="c-footer">
-  <div class="wrap">
-    <div class="signup">
-      <!--<p class="title -mono-font -center -white">-->
-      <!--<a class="sign-up-link" target="_blank"href="https://traseearth.squarespace.com/">-->
-      <!--Click here to receive news updates about TRASE-->
-      <!--</a>-->
-      <!--</p>-->
-    </div>
-    <div class="contain-logos-footer">
-      <div class="contain-logos">
-        <h4 class="title -mono-font">Trase is a partnership between</h4>
-        <ul class="logo-list">
-          <li class="logo-item sei-logo">
-            <a href="https://sei-international.org/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/partnership/s-e-i@2x.png" alt="Stockholm Environment Institute">
-            </a>
-          </li>
-          <li class="logo-item gcp-logo">
-            <a href="http://globalcanopy.org/" target="_blank" rel="noopener noreferrer">
-              <img class="gcp-img" src="/images/footer/partnership/g-c-p@2x.png" alt="Global Canopy Programme">
-            </a>
-          </li>
-        </ul>
-      </div>
-
-      <div class="contain-logos">
-        <h4 class="title -mono-font">In collaboration with</h4>
-        <ul class="logo-list">
-          <li class="logo-item vizzuality-logo">
-            <a href="http://www.vizzuality.com/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/collaboration/vizzuality@2x.png" alt="Vizzuality">
-            </a>
-          </li>
-          <li class="logo-item efi-logo">
-            <a href="http://www.efi.int/portal/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/collaboration/e-f-i@2x.png" alt="The European Forest Institute">
-            </a>
-          </li>
-        </ul>
-        <div class="contain-text">
-          <p class="title -mono-font">and many other organizations and individuals.</p>
-        </div>
-      </div>
-
-      <div class="contain-logos">
-        <h4 class="title -mono-font">Donors</h4>
-        <ul class="logo-list -donors">
-          <li class="logo-item wwf-logo">
-            <a href="https://www.worldwildlife.org/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/wwf@2x.png" alt="WWF">
-            </a>
-          </li>
-          <li class="logo-item gef-logo">
-            <a href="https://www.thegef.org/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/gef@2x.png" alt="The Global Environment Facility">
-            </a>
-          </li>
-          <li class="logo-item eustars-logo">
-            <a href="https://ec.europa.eu/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/eu-stars-33@2x.png" alt="Europe">
-            </a>
-          </li>
-        </ul>
-        <ul class="logo-list -donors">
-          <li class="logo-item conservancy-logo">
-            <a href="https://www.nature.org/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/nature-conservancy@2x.png" alt="The Nature Conservancy">
-            </a>
-          </li>
-          <li class="logo-item sida-logo">
-            <a href="http://www.sida.se/English/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/s-i-d-a-logo@2x.png"
-                   alt="The Swedish International Development Cooperation Agency">
-            </a>
-          </li>
-        </ul>
-        <ul class="logo-list -donors">
-          <li class="logo-item moore-logo">
-            <a href="https://www.moore.org/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/moore@2x.png" alt="Gordon and Betty Moore Foundation">
-            </a>
-          </li>
-          <li class="logo-item ukaidsv-logo">
-            <a href="https://www.gov.uk/government/organisations/department-for-international-development"
-               target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/u-kaidsvg@2x.png" alt="Department for International Development">
-            </a>
-          </li>
-          <li class="logo-item formas-logo">
-            <a href="http://www.formas.se/en/" target="_blank" rel="noopener noreferrer">
-              <img src="/images/footer/donors/formas@2x.png" alt="Formas">
-            </a>
-          </li>
-        </ul>
-      </div>
-    </div>
-    <div class="contain-link-list">
-      <ul class="links-list">
-        <li class="link-item"><a class="title -mono-font" href="/terms-of-use">Terms of Service</a></li>
-        <li class="separator"> ·</li>
-        <li class="link-item"><a class="title -mono-font" href="mailto:info@trase.earth"> Contact us</a></li>
-      </ul>
-    </div>
-  </div>
-</footer>
